refactor code in app-navigator
